moves resources into a module
