+++ src/engine.rs
@@ -3,8 +3,7 @@ use crate::blocker::{Blocker, BlockerError, BlockerOptions, BlockerResult};
 use crate::lists::parse_filters;
 use crate::request::Request;
 use crate::filters::network::NetworkFilter;
-use crate::resources::Resource;
-use crate::redirect_resources::{RedirectResource};
+use crate::resources::{Resource, RedirectResource};
 use rmps;
 use flate2::write::GzEncoder;
 use flate2::read::GzDecoder;
