+++ src/resources/resource_assembler.rs
